Improve structure and documentation of FilterList connector
