+++ Sources/InstantSearchCore/FilterList/Connector/NumericFilterListConnector.swift
@@ -1,5 +1,5 @@
 //
-//  FacetFilterListConnector.swift
+//  NumericFilterListConnector.swift
 //  InstantSearchCore
 //
 //  Created by Vladislav Fitc on 26/11/2019.
@@ -8,26 +8,26 @@
 
 import Foundation
 
-/// Filtering component that displays any kind of facet filters and lets the user refine the search results by selecting them.
-public typealias FacetFilterListConnector = FilterListConnector<Filter.Facet>
+/// Filtering component that displays any kind of numeric filters and lets the user refine the search results by selecting them.
+public typealias NumericFilterListConnector = FilterListConnector<Filter.Numeric>
 
-public extension FacetFilterListConnector {
-      
+public extension NumericFilterListConnector {
+
   /**
   - Parameters:
-    - facetFilters: List of facet filters to display.
+    - numericFilters: List of numeric filters to display.
     - selectionMode: Whether the list can have `single` or `multiple` selections.
-    - filterState: The filter state  that will hold your filters.
+    - filterState: Filter state holding your filters
     - operator: Whether we apply an `and` or `or` behavior to the filters in the filter state.
     - groupName: Filter group name
   */
-  convenience init(facetFilters: [InstantSearchCore.Filter.Facet] = [],
-                   selectionMode: SelectionMode = .multiple,
+  convenience init(numericFilters: [InstantSearchCore.Filter.Numeric] = [],
+                   selectionMode: SelectionMode = .single,
                    filterState: FilterState,
                    `operator`: RefinementOperator,
                    groupName: String) {
-    let interactor = FacetFilterListInteractor(items: facetFilters,
-                                               selectionMode: selectionMode)
+    let interactor = NumericFilterListInteractor(items: numericFilters,
+                                                 selectionMode: selectionMode)
     self.init(filterState: filterState,
               interactor: interactor,
               operator: `operator`,
@@ -36,20 +36,20 @@ public extension FacetFilterListConnector {
   
   /**
   - Parameters:
-    - facetFilters: List of facet filters to display.
+    - numericFilters: List of numeric filters to display.
     - selectionMode: Whether the list can have `single` or `multiple` selections.
-    - filterState: The filter state  that will hold your filters.
+    - filterState: Filter state holding your filters
     - operator: Whether we apply an `and` or `or` behavior to the filters in the filter state.
     - groupName: Filter group name
-    - controller: Controller interfacing with a filter list view
+    - controller: Controller interfacing with a concrete filter list view
   */
-  convenience init<Controller: SelectableListController>(facetFilters: [InstantSearchCore.Filter.Facet] = [],
-                                                         selectionMode: SelectionMode = .multiple,
+  convenience init<Controller: SelectableListController>(numericFilters: [InstantSearchCore.Filter.Numeric] = [],
+                                                         selectionMode: SelectionMode = .single,
                                                          filterState: FilterState,
                                                          `operator`: RefinementOperator,
                                                          groupName: String,
-                                                         controller: Controller) where Controller.Item == InstantSearchCore.Filter.Facet {
-    let interactor = FacetFilterListInteractor(items: facetFilters,
+                                                         controller: Controller) where Controller.Item == InstantSearchCore.Filter.Numeric {
+    let interactor = NumericFilterListInteractor(items: numericFilters,
                                                selectionMode: selectionMode)
     self.init(filterState: filterState,
               interactor: interactor,
